Add linters, reorganize code structure, and move to ES6
