+++ assets/css/style.css
@@ -83,6 +83,8 @@ body {
   margin: 0 10px 0 0;
   text-decoration: none;
   -webkit-transition: color 0.2s;
+  -moz-transition: color 0.2s;
+  -o-transition: color 0.2s;
 }
 
 .tag::before {
