+++ test/csslint.spec.js
@@ -1,0 +1,4 @@
+/**
+ * CSS Linting
+ */
+require('mocha-csslint')(['assets/css']);
