+++ test/eslint.spec.js
@@ -1,0 +1,18 @@
+/**
+ * JS Linting
+ */
+
+let colors = require('mocha/lib/reporters/base').colors;
+colors['pass'] = '1;36';
+colors['error stack'] = '1;36';
+colors['fast'] = '1;36';
+colors['light'] = '1;36';
+colors['diff gutter'] = '1;36';
+
+require('mocha-eslint')([
+  'config',
+  'models',
+  'routes',
+  'test',
+  'app/app.js'
+]);
